+++ faddr/__init__.py
@@ -8,13 +8,5 @@ from rich.console import Console
 
 __version__ = "0.4.0-alpha.1"
 
-if os.getenv("FADDR_DEBUG") or any(arg in sys.argv for arg in ("-D", "--debug")):
-    LOG_LEVEL = "DEBUG"
-else:
-    LOG_LEVEL = "INFO"
-
-logger.remove()
-logger.add(sys.stdout, level=LOG_LEVEL)
-
 # Setup rich console for pretty printing
 console = Console()