+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - REST API v0 using FastAPI and uvicorn
+- `RepoList` class to aggregate repos
 
 ## [0.3.1] - 2022-04-19